backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
supabase==2.5.0
httpx==0.27.0
slowapi==0.1.9
pydantic-settings==2.2.1
pydantic==2.7.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.2
xlsxwriter==3.2.0
pandas==2.2.2
